requirements.txt
fastapi
uvicorn
bcrypt
python-jose[cryptography]
httpx
pydantic[email]
